+++ src/river/River.java
@@ -1,4 +1,8 @@
 package river;
+import util.Util.*;
+
+import static util.Util.onetofive;
+import static util.Util.timeString;
 
 public class River {
     private int chance;
@@ -6,7 +10,7 @@ public class River {
 
 
     public void reroll(){
-        chance=Util.onetofive();
+        chance= onetofive();
 
     }
 
@@ -19,16 +23,16 @@ public class River {
                 kidEvent();
                 break;
             case 2:
-                Util.timeString("주인공이 수영을 합니다",  100);
+                timeString("주인공이 수영을 합니다");
                 break;
             case 3:
-                Util.timeString("아무일도 없었다 그만 돌아가자",100);
+                timeString("아무일도 없었다 그만 돌아가자");
                 break;
             case 4:
-                Util.timeString("강가에서 서아를 만났다 ...\n안녕..",  100);
+                timeString("강가에서 서아를 만났다 ...\n안녕..");
                 break;
             case 5:
-                Util.timeString("오늘은 낚시를 했다",  300);
+                timeString("오늘은 낚시를 했다");
                 break;
             default:
                 System.out.println("디폴트");
@@ -38,7 +42,7 @@ public class River {
     }
 
     public void kidEvent(){
-        Util.timeString("냇가에서 어떤 꼬마가 허우적 거리고 있다",  100);
+        timeString("냇가에서 어떤 꼬마가 허우적 거리고 있다");
     }
